Dependency and small config.h cleanup
Got it compiling on a brand new dev install, needed to swap out some dependencies for compatible ones and changes of repo name.

Also did a first pass at sorting variables into spaces that own them, with the idea that we can wrap abstractions around these and making these areas easier to extend and modify in the future.

Dependency and small config.h cleanup

Got it compiling on a brand new dev install, needed to swap out some dependencies for compatible ones and changes of repo name.

Also did a first pass at sorting variables into spaces that own them, with the idea that we can wrap abstractions around these and making these areas easier to extend and modify in the future.